requirements.txt
streamlit==1.21.0
numpy==1.24.0
Pillow==9.5.0